requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Django==2.2
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
flake8==3.8.3
idna==2.10
lxml==4.5.2
mccabe==0.6.1
mypy==0.782
mypy-extensions==0.4.3
numpy==1.19.1
pandas==1.1.0
pathspec==0.8.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
